remove schema usage in migration sql
using schemas adds complexity in the rust code that has to work with
both pg and sqlite. Since schemas don't exist in sqlite (for a single
database), a rust chunk will have to check if it's working with a pg
connection and a sqlite connection and depending on this the query will
execute on table `fastn.<table_name>` (for pg) or `fastn_table_name`
(for sqlite). We can get rid of this by simply not using schemas in pg.