Split `Request::session_id()` into `session_id`...
... and `preview_session_id()`

The intent to add the `Request::session_id` method was to check if the
request is to render a preview (this is used on fifthtry) or it's a
normal request. This method has since been used in other places (eg. in
the user_details processor).

To separate these concerns of reading authenticated user data and
rendering previews, the old method is now modified to read only the
secure same site cookie (`fastn-sid`). And, the preview related methods
now use `Request::preview_session_id()` which only reads the cross-site
`fastn-preview` cookie.