Remove markdown parsing from frontend
[marked](https://github.com/markedjs/marked/) was used to parse markdown
for each text node in the browser which was making things very slow.

This commit moves markdown parsing to the backend using the `comrak`
crate.

Some cleanups may be required that removes the marked.js files
completely from the codebase but it'll happen later once I get this
changes reviewed.