we have to go straight to type-checker now
earlier we were going to go from p1 to ast0, but we can skip that step
and go straight to type checker.

made more breaking changes. we have a `fastn_p1::Document` now, and
removed the `fastn_p1::Item` and using fields.

p1 is now going to be responsible for parsing import statements as well
as then we can classify the rest of sections as either content or
definition. creating one list of sections, and then splitting it into
imports, definitiosn, content is kind of useless.

also extracted out comments and errors because we can ignore all
comments most of the time, so why complicate our rest of code with
comment ignore logic, and also errors are only looked at at the end, so
we do not need to keep them also interleaved with definitions and
comment.