Optimize images on push to `main` and create a PR with changes
Images are now optimized when commits are pushed to the `main` branch.
This way we can run this on PRs from forks too.

A new PR with optimized version of the images wil be created instead of
directly commiting to `main`. This way we have a chance to review things
before things go in `main`.