instead of string-interner, lets use arcstr
Ideally I would have liked to use rcstr, we do not need arc, rc is good
enough, but the rcstr crate, while it exists, i do not feel confident
about it.

we are not going to stop using string-interner. we will use
string-interner for IDs like package name, module name, symbol name etc,
as interned-strings are faster to compare than arcstr, which is
internally still a string, and equality is o(k) where k is length of
shorter string. with intern the equality is o(1) as it compares just the
pointers (u32 actually).