not keeping fastn_unresolved::Documents.imports at all
since we can be in two states, before import desugaring, and after
import desugaring with the current design, where we create document wit
imports, and then desugar in two separate steps.

this is wrong because we can accidentally forget to do that, so lets
desugar import before finishing parsing at all, so we never are left in
improper state.