Refactor: custom `fastn_core::http::Response` for `fastn_core::serve`
`actix_web::HttpResponse` aka `fastn_core::http::Response` is not `Send`
which makes it impossible to share it across threads. A direct downside
of this is that we can't use a function which uses this type in a
`tokio::spawn` because it required types to implement `Send`.

A custom `fastn_core::http::Response` is added which is `Send` and used
by the `fastn_core::serve` function. A `fastn_core::serve_actix` is also
introduced which is a direct replacement for the old
`fastn_core::serve`.

This will allow us to, for example, use `fastn_core::serve` in a
background task to update the cache.