We do not store ModuleAlias but Module
In earlier design we did not have a concept of a Module, meaning in
Definition we did not have a clause for Module. We had module alias and
symbol alias instead.

The aliases are not things, where as module is a thing. This commit
fixes that mistake.

Further, we are going to add alias as a id-arena map on every
definition, which will contain the merged value of aliases from globals,
auto imports and imports designed in current module.