Feat: Implicitly load `.wasm` modules from mounted apps
```ftd
-- fastn.app: Auth App
mount-point: /-/auth/
package: lets-auth.fifthtry.site
```

For the above configuration in a FASTN.ftd file. We allow the mounted
app package to place some `.wasm` modules in its root directory.

So if the request is coming for: `/-/auth/api/*` and there is a
"api.wasm" in the root of `lets-auth.fifthtry.site` then the request
will be handled as if there was a wasm endpoint configured:

```ftd
-- fastn.url-mappings:

/-/auth/api/* -> wasm+proxy://.packages/lets-auth.fifthtry.site/api.wasm/*
```

That is, a request for `/-/auth/api/*`, will be handled by the
`lets-auth.fifthtry.site/api.wasm` if it exists.

If the `.wasm` module does not exist then the typical flow of loading
app's .ftd file is followed. This means that precedence is given to
.wasm if it exists.