Rename `ftd.app-path` -> `ftd.app-url`
`fastn` touches both the file system (through its fs routing) and the
urls (since it's a webserver, duh). This might create a situation where
people think `ftd.app-path` is something related to fs but it's really
constructing the path relative to your website url. For this reason,
this has been renamed to `ftd.app-url` instead.