Fix: Construct correct path when `fastn.app`'s mountpoint is '/'
If a `fastn.app` is mounted at `/` (root) then requests coming to the
root url (just the hostname in the browser) lead to construction of an
incorrect path.

For example:

```ftd

-- fastn.app: Test
package: lets-talk.fifthtry.site
mount-point: /
```

With the above setup, a request to `/`, would create a "package
sanitized path": `-/lets-talk.fifthtry.site//`.

The trailing double-slash (//) does not cause any problems as it's
resolved correctly when used with a real filesystem (fastn_ds::Path,
based on camino::Path, is able to fix this). However, this does become a
problem when the paths are not resolved. It's better to never construct
a path like this so it never relies on fs to do the right thing.