Feat: Show error when a thing is defined using an alias name
```ftd
-- import: my-package/my-module

-- my-module:
```

The above code snippet makes the server go into an infinite loop. This
shouldn't be allowed as we're trying to directly render a module. In
most cases `-- my-module.a-component:` is what people want.

```ftd
-- import: my-package/something

-- string something: some value

-- ftd.text: $something
```

The above code snippet will also throw the same error. Trying to define
a thing (variable in this case) with the same name as an alias
(`something` import alias in this case) is not allowed.