removing fastn_user and _session migrations
We are moving the migrations to [lets-auth][1], as that provides every
other auth related functionality.

We are getting rid of auth-provider and instead of going to provide
integrated auth solution, this is to reduce the complexity and simplify
things for now, if some pattern emerges in future we will think about it
then. for now simple auth package is enough of a framework we want to
support.

[1]: https://github.com/fifthtry-community/lets-auth/pull/22