Fix: Consider `sitemap` items when resolving imports
```ftd
;; guide/install.ftd
-- string test: some val
-- ftd.text: $test
```

```ftd
;; FASTN.ftd
...

-- fastn.sitemap:

\# Installation: /install/
  document: guide/install.ftd
```

This will fail to build, reason:

fastn is trying to resolve `$test` in `guide/install.ftd` and in doing
so, it's resolving the current module with id: `/install/` but this
fails as the id translates to a `install.ftd` (or `/install/index.ftd`)
in project dir. The correct id should be `guide/install/` (document
attributed in sitemap definition).

This is fixed by looking into sitemp entries if available and changing
the id before trying to do `package.resolve_by_id`.