Build static fastn for x86_64-linux-gnu
The CI is setup to build static `fastn` binary. This is done by
statically linking with the glibc on linux.

This is tested on NixOS with the `pkgs.glibc.static` package. From
stackoverflow[1], it's equivalent on ubuntu is `libc6-dev` so this
should work on github CI too.

An attempt to compile using `musl-gcc` was also made but it did not work
because crates `rusqlite` (because libsqlite-sys) and `quickjs` failed
to link with musl's libc.