feat: Support error handling in tuple variant of ftd.submit_form
In an ftd function body, you can use the tuple variant of
`ftd.submit_form` method. This does not require you to have a
`ftd.string-field` instance.

```ftd
ftd.submit_form(
    action_url,
    ("concept-url", concept_url),
    ("status", status, error)
)
```

The third element of the tuple is a mutable variable defined as:

```ftd
-- optional string $status-error:
```

Note that the use of `optional` is necessary as using just the `string`
type would issue warning and will not get updated from js.