fix(http processor): Handle wasm modules in mountpoints of apps
We check if the url is part of a mounted app in the current package. If
it is, then we try to create a wasm+proxy:// url for implicit wasm
modules present in the app's directory. If the implicit wasm module does
**not** exists, we simply bail by creating a full url (prefixed with
http) and let the fastn server logic handle it when the http processors
calls the full url.