fix(http processor): Handle package local relative urls
Instead of returning an error in case of no resolution of the url. We
simply return the full url (path prefixed with scheme+host) in hopes
that fastn server logic will handle it.

This is useful in situations like the following:

```ftd
;; index.ftd
-- some-record-name d:
$processor$: processors.http
url: /local/
```

```ftd
;; local.ftd

-- ftd.json
id: 1
name: Someone
```

The resolved will convert url `/local/` -> `http://127.0.0.1:800/local/`
(depends on running env) which is where the http request from the
processor will be sent and fastn server will realise that it's a local
file and will return the correct json response.