fix: Actually working system packages!
This is fixed by prefixing the auto-import with `inherited-` if:

The auto-import is for the dependency that has a `provided-via`
attribute which has the value of the path to a local ftd file that
overrides the contents of this dependency.

After the prefix is added. We check for this prefix while resolving
imports, if the imported module starts with `inherited-` we instead read
it's `provided-via` file path and load its contents instead. This way
even the third party dependencies end up importing this `imported-`
module by sourcing the `provided-via` path.

The users don't have to know about this `inherited-` prefixing as it is
done by the framework to the auto imports only.

Explicitly importing the `provided-via` package by its name works as
usual and will import the un-overriden original variant of the
dependency. This is useful in case where you want to override a few
components in your `provided-via` file but re-export original symbols by
importing the original dependency.

fixes: #2139