fix(http processor): use string contents when json is just string
Calling `serde_json::to_string` on a `serde_json::Value::String`
produces a quoted string which is a valid json value but not what we
want to put inside url query params. To handle this, we check if the
value is a String and use the inner string value if it is. In all other
situations we want a safely escaped json value that we can decode on the
other side (ft_sdk).