fix: Server render meta tags if request is from a bot
A bot for example is Linkeding | facebook | discord etc. We check if the
request is coming from them, then we take the generated js and execute
it on the server (ssr) to create those meta tags and then inject them
into the output html.

Normal browser request don't take the perf hit of running js on the
server.

This requires no change in user code. Existing use of `ftd#document`
should just work.