feat: Tauri setup and some refactor
Break the fastn crate into a lib and a main binary. Most of the
functions in main.rs have been moved into lib.rs.

The tauir.rs deals with tauri specific stuff and contains the entrypoint
for mobile build of fastn.

The assets and configs are generated using `cargo tauri init`