fix: Avoid infinite loop
Avoids infinite loop for the following code snippet:

```ftd
-- test:
a: Hello

-- component test:
string a:

-- ftd.text: Test
classes: $test.a

-- end: test
```

`classes` of `ftd.text` component takes a list of string but ftd has a
feature that turns T -> List<T> if expected kind is List<T> and found
kind is T.

This works for single literals, like:

```
-- ftd.text: something
classes: one
```

The string literal `one` is converted to List<String> because `classes`
expects that.

This, for some reason, does not work when we use a variable reference
instead of literals. This changes fixes this case.

NOTE: This change does not support mixing of two styles, you can't do:

```ftd
-- test:
a: two

-- component test:
string a:

-- ftd.text: Test
classes: one, $test.a ;; SEE THIS!

-- end: test
```

This still results in a Cycle Detection error:

```
failed to parse FoundCycle { message: \"test#test:4 => test#test:4 => test#test:4\", line_number: 4 }
```

Simply removing this match branch and failing with a "expected_kind and
found_kind mismatch" error is not an option here as it'll break so many
existing packages in the world.