feat: Release build for android
The debug build does not work because fastn-core crashes with stack
overflow. Things work fine in release builds.

The catch is, tauri is forcing to sign the release apk before it can run
it in the emulator. I have create a dev signing script for this. This
will have to change and we have to create a signing key when releasing
it through github releases.