fix: Avoid caching for android
The cache_it function depends on a cache dir which it can't fetch on
android. The path to the cache dir has to come from tauri, this requires
changing params of some functions so I'm simply avoiding it for now.

FWIW we don't ever read from this cache (the code is commented out) so
caching it is just extra work with no benefits. I could be wrong tho and
we might be reading from this cache in some other place but it does not
seem like it from my quick inspection.