feat: /contribute/ page
Mention our Open Collective on this page. Link to /contribute/ in header
and footer automatically appear. We should also add a section on the
landing page but I am not doing it right now (I feel the website source
code can be made simpler before I do this).

Others wasy this page will be useful:
- Link to code contribution guide (github or a page on our website)