feat: A barebones SMTP server
This commit starts the fastn-mail crate. This will handle the SMTP/IMAP
and related email protcols. It'll also contain the bridge code so that
email clients can talk to this but the email get delivered over the iroh
network.

A simple SMTP impl is done, a lot of the things are left:
- support full SMTP proto
- upgrade to encrypted connection STARTTLS. Some client might depend on
  this.
- Actually deliver emails
- IMAP impl so that clients can fetch emails

The current impl is able to handle the tests/send-mail.py script. This
was just an experimentation to learn basics of SMTP protcol.