feat: Support for catch-all 404/500 pages
If a fastn package contains a 404.ftd file then visiting any 404 route
(a route that does not exist) will render this 404.ftd file.

For any other fastn error (error in syntax, calls to http processor
failing etc), the 500.ftd in current fastn package will be rendered.

when env DEBUG is set, the original error with a nicer message is shown
instead to give the developer some more context. In production, the
`--trace` output should be considered around these 404/500 requests to
know about the cause.